requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.2
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
Pillow==9.5.0
pyparsing==3.1.1
pypng==0.20220715.0
python-dateutil==2.8.2
qrcode==7.4.2
six==1.16.0
typing_extensions==4.9.0
